requirements.txt
streamlit>=1.32
langgraph>=0.2
langchain>=0.2
langchain-community>=0.2
langchain-google-genai>=1.0
tavily-python>=0.3.3
python-dotenv>=1.0
langsmith>=0.1.77
